Created hang.js: Serves as hangman clone based on the spy from Team Fortress 2.
User starts the hangman game when he types s!hang. The user types single
letters into the chat to make guesses. The word is split into an array
with each char it's own index. As the user guesses correct letters, the
corresponding indexes in the array are turned into blank chars ('').

When the array is full of only empty chars, the user has successfully
guessed all of the letters and wins the game.

On the other hand, if the user fails to guess a correct letter 6 times,
then they fail the game.

The game is shown on a Discord Embed object with :sandwich: representing
the guessable letters and :blue_square: representing spaces. The Embed
also features an image that changes to a gif if the player loses the
game.